Adding support to not use SSL when using SASL
According to the documentation it is possible to not force the use of SSL
when using SASL authentication.
Since in some cases the Kafka cluster is inside the same network without
external access, using SSL is a bit of overkill.

Added the possibility to not use SSL for the Fluentd plugin

Signed-off-by: Ricardo Verhaeg <ricardo@arquivei.com.br>