out_rdkafka2: Add a new parameter `max_enqueue_bytes_per_second`
When huge amount of messages are attempted to send, it causes excessive
load on both Fluentd and Kafka. Such situation might occur after down
time of Fluentd or Kafka.
Although it can be controlled by combination of following parameters, it
isn't intuitive and might be hard to understand:

* Fluentd:
  * `flush_thread_burst_interval`
  * `buffer_queue_limit`
  * `buffer_chunk_limit`
* fluent-plugin-kafka
  * `enqueue_retry_backoff`
  * `max_enqueue_retries`
* librdkafka
  * `queue.buffering.max.kbytes`
  * `queue.buffering.max.messages`
  * `queue.buffering.max.ms`

This commit enable to control the enqueueing rate by a simple parameter.

Signed-off-by: Takuro Ashie <ashie@clear-code.com>