feat(out_rdkafka2): `sasl.username` and `sasl.password` is now configurable via plugin config
While we were able to work around this by providing credentials via
`rdkafka_options` configuration parameter, this is not ideal since
`rdkafka_options` is not marked as a `secret` configuration parameter,
and thus the credentials will be logged to `stdout` during Fluentd start
up.

This commit makes it such that we are able to leverage the existing
`password` parameter that is already marked as secret.

Signed-off-by: Ray Tung <code@raytung.net>