in_kafka_group: send heartbeat if necessary
We've spotted that processing a batch of Kafka messages in fluentd
sometimes takes a long time and because of that Kafka consumer groups go
into a constant rebalancing state. Fluentd lacks observability to understand
what plugin is taking a long time so adding this as a stop gap solution
for the moment. It's not trivial to add duration histograms to Fluentd
so this is a stop-gap solution for now.

Kudos to @marijus-ravickas for the idea. More details available at
https://www.marionzualo.com/2019/08/10/always-manually-send-heartbeats-when-processing-message-batches-in-ruby-kafka/.

Signed-off-by: Giedrius Statkevičius <giedrius.statkevicius@vinted.com>