GitHub issue: Set waiting-for-triage by default
We might not check issues so often recent days, so it is inappropriate
to close stale issues that aren't checked by any maintainers.
This commit will add waiting-for-triage label to all new issues and
won't close them by default. In addition, issues that are assigned or
set milestone are also exempted.

Signed-off-by: Takuro Ashie <ashie@clear-code.com>